CSS tweaks: quotes, notes, and iPhone-specifics
Make the blockquote and sidenote styles a little less ugly and a little
more cohesive -- quotes are in blue and notes are in pink.  Also, remove
most of the iPhone-specific styling as it just gets in the way.  The
site is perfectly readable in its default style if you turn the phone
sideways.  I have left a bit of iPhone-specific styling in there, just
to make the header a bit bigger, but other than that it looks good
as-is.