Split config into separate modules
This will make it easier to describe each portion of the configuration
in a different blog post.

While I was doing this, I turned on -Wall and fixed the warnings that
came up.  Bit off to have those changes mixed in with the splitting
apart of modules, but reverse-engineering my changes into discrete
commits is too much effort at this point so I'm just going to check it
in and have done with it.