Move images into directory based on post name
The idea here is that instead of storing images in a folder separate
from the posts, they end up in a subfolder of posts named after the
post itself.  This way, I can make the post itself the index.html of
that folder, and I get nicer post urls and images can be referenced
locally.