Update for the Hakyll 4.9.3.0
Most of the required updates are because Hakyll has switched to using
JSON (via aeson) for its metadata, rather than a simple Map of Strings.
parseTime has also been deprecated in favour of parseTimeM.  Finally,
I had to do some faffing about with the cabal file to get it to build,
including working with a local clone of Hakyll that had some dependencies
tweaked, so this is probably not ready to check in.