Add support for 'prefers-color-scheme'
The site will now default to dark mode when the system is in dark mode
and light mode when it is in light mode.  In dark mode, URLs can be
suffixed with '?light' to make them light; in light mode, they can be
suffixed with '?dark' to make them dark.  The button suffixes
appropriately.